Rename state.js to BridgeState.js
