+++ src/BridgeState.js
